MD5-Hash: Optimization in Endianness Change
-Previously the byte array was converted to an int array and was then
endian swapped afterwards.

Because the conversion from byte array to int array already requires us
to index the bytes and shift them into a 4 byte value, a simpler option
would be to change the shift values to set it to little endian
immediately.

This saves a lot of work doing it afterwards. A comment has been added
to ensure this change is obvious. Readability should stay roughly the
same.