MD5-Hash: Correction to Writing Message Size
-Previously the message size was being written backwards and to the
wrong offset. Because it was backwards and near the correct offset, it
appeared to be correct because the byte that eclipsed the correct spot
was the most important byte.

This meant that when the size was less than or equal to the value of a
single byte the algorithm would work as intended. But when the message
size was multiple bytes it would be written incorrectly.

This was solved by reversing the value being written and adjusting the
offset to adjust for it.

-Additionally the mask was updated to match as the value is a 32 bit
value, not 8 bit. Even if it was 8 bit, the mask would have needed to be
0xFF00, not 0xF000. This was a typo and a lapse in thought.