MD5-Hash: Big Optimization in rotateLeft/rotateRight Functions
-Previously there were two rotateRight and two rotateLeft functions. One
accepted the value and the shift amount and called the function that
does the actual rotation as many times as the value of the shiftAmount
argument. It simply looped and called a function that does a single
rotation.

This is bad as it means there is both a loop constantly iterating and
constantly calling a function, returning a value, setting a value, and
repeating.

This function could occur in a single go by modifying the function that
does the rotate and modify the shifts in it to account for the shift
amount in the first place.

This allows for an RVO optimization as well as a much simpler and faster
function.