MD2: Now supports input larger than a single chunk
-Previously the implementation was not taking in account several chunks
of input. It was calculating the hash based on the whole padded input
(which with previous tests was only a single chunk)

Now the padded message is buffered into an array which the checksums are
calculated from until the whole message has been calculated.

This now calculates an arbitrary amount of chunks as confirmed by a new
test.