MD2: Loop Optimization for Padding
-Optimized padding. Instead of a single loop with an if else constantly
re-checking the condition and deciding which operation to do it has been
split into two loops that do their specific purposes.

One loop copies the data from the original message to the larger padded
message. The second writes the padding data that follows the original
message.

This saves processing power/time and is more clean to read.