MD-Hash: Fixed Padding for MD4 Algorithm
-MD4 Padding failed when the message length was 9 bytes away from the
nearest multiple of 64. This is primarily because it decided how many
bytes needed be padded from a number that was already larger than the
actual message. So a window of error was created in which this larger
number didn't need to be padded while the actual message did, therefore
it would not pad at all.

This is now fixed because this check is done with the actual message
length instead. Therefore it will always calculate how many bytes of
padding are needed for the message itself.

-Although the above problem is fixed, it leaves another issue that the
implementation described above attempted to fix. The reason
message.Length+9 was being compared was because it was attempting to
ensure the padding would be at least 9 bytes. The implementation was
flawed, but the issue needed to be fixed regardless.

The current answer to this is that now the message will be padded to the
next multiple of 64 no matter what the message length is. This could
result in a situation where the message has been padded, but does not
have the minumum of 9 bytes of padding.

When this is the case, an additional 64 bytes of padding will be added
on top of the previous padding. This should result in the correct
padding regardless of the message length.