SHA-Hash: Fixed bug indexing sha1Table
Previously the Sha1 table was composed of  the same 4 variables written
80 times so that for each i there was a copy of the correct variable to
be indexed from the array directly.

Afterwards the table was simplified down to the 4 single variables. When
this occurred the code written to index the proper variable did not get
tested appropriately. Instead an old version had been run.

After rebuilding and testing it was found that this code was not
correct. It has now been corrected to appropritately index one of the 4
variables.