FNV-Hash: Correction to FNV0 Legacy Hash
-A mistake was made with the implementation of the fnv0 hash. The
difference between FNV0 and FNV1 is the initialization of the hash
variable. In FNV0 it is initialized to 0, where as in FNV1 there is a
specific offset it is set to.

My implementation had set the prime for FNV-0 to 0, which leads to
drastically different results since it's constantly being multiplied by
0 each step then having the data spliced in with an XOR operation.